+++ src/main/java/nextstep/auth/ui/AuthenticationPrincipalArgumentResolver.java
@@ -3,7 +3,7 @@ package nextstep.auth.ui;
 import nextstep.auth.AuthenticationException;
 import nextstep.auth.application.AuthManager;
 import nextstep.auth.application.TokenType;
-import nextstep.subway.member.domain.LoginMember;
+import nextstep.auth.domain.LoginMember;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
