+++ src/main/java/nextstep/subway/favorite/ui/FavoriteController.java
@@ -5,7 +5,7 @@ import nextstep.auth.ui.AuthenticationPrincipal;
 import nextstep.subway.favorite.application.FavoriteService;
 import nextstep.subway.favorite.application.dto.FavoriteRequest;
 import nextstep.subway.favorite.application.dto.FavoriteResponse;
-import nextstep.subway.member.domain.LoginMember;
+import nextstep.auth.domain.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
