+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -4,7 +4,7 @@ import nextstep.auth.ui.AuthenticationPrincipal;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.application.dto.MemberRequest;
 import nextstep.subway.member.application.dto.MemberResponse;
-import nextstep.subway.member.domain.LoginMember;
+import nextstep.auth.domain.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
